Improved CSS string parsing.
Switched CSSTidy's string parsing algorithm to use a stack to manage nested strings (which includes parenthetical statements, which can be nested infinitely), rather than a single str_in_str flag. This fixes the incorrect handling of newlines inside of parentheses, which are valid but were being treated as newlines inside of quoted strings.